+++ translations/be.py
@@ -185,6 +185,16 @@ def translateNewVmBE(window):
 
     # Fourth page
     window.lbl_vga.setText("VGA") # VGA
+
+    i = 0
+
+    while i < window.cb_vga.count():
+        if window.cb_vga.itemText(i) == "Let QEMU decide" or window.cb_vga.itemText(i) == "QEMU überlassen":
+            window.cb_vga.setItemText(i, "Няхай QEMU вырашае") # Let QEMU decide
+            break
+
+        i += 1
+
     window.lbl_net.setText("Сетка") # Network
     window.lbl_mouse.setText("Мыш") # Mouse