+++ CHANGELOG.md
@@ -56,6 +56,8 @@
 ## Changes compared to v2.1.0.5704_b1
 
 - Logging has been applied to starting VMs.
+- For existing translations, the following has been set up:
+    - VGA combobox "Let QEMU decide".
 
 ## Known issues